+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- The `get_by_full_name()` function has two new optional arguments `instance_of` and `subclass_of`, which may be 
+  used to ensure correct return types.
+
 ### Fixed
 - Print parent of inner class in `misc.tname()` when `prefix_classname=True`.